Important security update!
The documentation for `realmPwdAlg` on the Progress KnowledgeBase was woefully inadequte and led to us making a large error in interpretation of what that option does. It affects the link between Spring and the AppServer, not the connection from the client to the server. There is no way to use digest auth for the outward facing connection, so you must **always** use SSL to protect your users' credentials!